+++ Yutai.ArcGIS.Carto/UI/RasterClassifiedRenderPage.cs
@@ -636,7 +636,7 @@ namespace Yutai.ArcGIS.Carto.UI
                 int num2;
                 int num3;
                 int num4;
-                int rGB = icolor_0.RGB;
+                uint rGB =(uint) icolor_0.RGB;
                 this.method_6(rGB, out num2, out num3, out num4);
                 colorEdit_0.Color = Color.FromArgb(icolor_0.Transparency, num2, num3, num4);
             }
